+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.0].define(version: 2026_09_04_120000) do
+ActiveRecord::Schema[8.0].define(version: 2026_09_09_151432) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "fuzzystrmatch"
   enable_extension "pg_catalog.plpgsql"
@@ -1103,6 +1103,7 @@ ActiveRecord::Schema[8.0].define(version: 2026_09_04_120000) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.bigint "created_by_id"
+    t.index ["board_id", "team_id"], name: "index_team_boards_on_board_and_team", unique: true
     t.index ["board_id"], name: "index_team_boards_on_board_id"
     t.index ["team_id"], name: "index_team_boards_on_team_id"
   end